Move instantiation into parametrization
