+++ nevergrad/instrumentation/__init__.py
@@ -3,11 +3,11 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .instantiate import register_file_type, FolderFunction
 from . import variables
 from . import variables as var
 from .multivariables import Instrumentation as Instrumentation
 from .multivariables import InstrumentedFunction as InstrumentedFunction
-from .utils import TemporaryDirectoryCopy, CommandFunction
+from nevergrad.parametrization.instantiate import FolderFunction
+from nevergrad.parametrization.utils import TemporaryDirectoryCopy, CommandFunction
 
 __all__ = ["Instrumentation", "var", "CommandFunction", "FolderFunction"]
