+++ nevergrad/parametrization/__init__.py
@@ -3,4 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+from nevergrad.parametrization.instantiate import FolderFunction as FolderFunction
+from nevergrad.parametrization.utils import TemporaryDirectoryCopy as TemporaryDirectoryCopy
+from nevergrad.parametrization.utils import CommandFunction as CommandFunction
 from . import parameter as parameter
